# Optimizes internal trace context related code, especially extra