--- a/brave/src/main/java/brave/internal/PropagationFields.java
+++ b/brave/src/main/java/brave/internal/PropagationFields.java
@@ -1,12 +1,11 @@
 package brave.internal;
 
 import brave.propagation.TraceContext;
-import java.util.List;
 import java.util.Map;
 
 /** Copy-on-write keeps propagation changes in a child context from affecting its parent */
 public abstract class PropagationFields {
-  private TraceContext context; // guarded by this
+  long traceId, spanId; // guarded by this
 
   /** Returns the value of the field with the specified key or null if not available */
   public abstract String get(String name);
@@ -20,14 +19,15 @@ public abstract class PropagationFields {
   public abstract Map<String, String> toMap();
 
   /** Fields are extracted before a context is created. We need to lazy set the context */
-  final boolean tryAssociate(TraceContext newContext) {
+  final boolean tryToClaim(long traceId, long spanId) {
     synchronized (this) {
-      if (context == null) {
-        context = newContext;
+      if (this.traceId == 0L) {
+        this.traceId = traceId;
+        this.spanId = spanId;
         return true;
       }
-      return context.traceId() == newContext.traceId()
-          && context.spanId() == newContext.spanId();
+      return this.traceId == traceId
+          && this.spanId == spanId;
     }
   }
 
@@ -36,28 +36,22 @@ public abstract class PropagationFields {
   }
 
   /** Returns the value of the field with the specified key or null if not available */
-  public static String get(TraceContext context, String name) {
+  public static String get(TraceContext context, String name,
+      Class<? extends PropagationFields> type) {
     if (context == null) throw new NullPointerException("context == null");
     if (name == null) throw new NullPointerException("name == null");
-    PropagationFields fields = find(context.extra());
+    PropagationFields fields = context.findExtra(type);
     return fields != null ? fields.get(name) : null;
   }
 
   /** Replaces the value of the field with the specified key, ignoring if not a permitted field */
-  public static void put(TraceContext context, String name, String value) {
+  public static void put(TraceContext context, String name, String value,
+      Class<? extends PropagationFields> type) {
     if (context == null) throw new NullPointerException("context == null");
     if (name == null) throw new NullPointerException("name == null");
     if (value == null) throw new NullPointerException("value == null");
-    PropagationFields fields = find(context.extra());
+    PropagationFields fields = context.findExtra(type);
     if (fields == null) return;
     fields.put(name, value);
-  }
-
-  public static PropagationFields find(List<Object> extra) {
-    for (int i = 0, length = extra.size(); i < length; i++) {
-      Object next = extra.get(i);
-      if (next instanceof PropagationFields) return (PropagationFields) next;
-    }
-    return null;
   }
 }