After some trial and error with benchmarking, I found
`MutableTraceContext` is not enough a win unless we change api, which we
can't until Brave 6. Meanwhile, this optimizes as much as possible,
notably eliminating as many arraylist creations (from extra) as
possible.

Alternative to #755