@adriancole 2018-08-18
here are the results, btw


before:

```
Benchmark                                                                                                       Mode      Cnt      Score      Error   Units
TracerBenchmarks.joinWithSpanInScope:joinWithSpanInScope·p0.999                                               sample              21.216              us/op
TracerBenchmarks.joinWithSpanInScope:·gc.alloc.rate.norm                                                      sample       15    791.518 ±    1.755    B/op
TracerBenchmarks.joinWithSpanInScope_extra:joinWithSpanInScope_extra·p0.999                                   sample              23.872              us/op
TracerBenchmarks.joinWithSpanInScope_extra:·gc.alloc.rate.norm                                                sample       15    782.750 ±    6.044    B/op
TracerBenchmarks.joinWithSpanInScope_unsampled:joinWithSpanInScope_unsampled·p0.999                           sample               1.100              us/op
TracerBenchmarks.joinWithSpanInScope_unsampled:·gc.alloc.rate.norm                                            sample       15     40.008 ±    0.005    B/op
TracerBenchmarks.joinWithSpanInScope_unsampled_extra:joinWithSpanInScope_unsampled_extra·p0.999               sample              31.611              us/op
TracerBenchmarks.joinWithSpanInScope_unsampled_extra:·gc.alloc.rate.norm                                      sample       15     40.055 ±    0.004    B/op
TracerBenchmarks.newChildWithSpanInScope:newChildWithSpanInScope·p0.999                                       sample              28.770              us/op
TracerBenchmarks.newChildWithSpanInScope:·gc.alloc.rate.norm                                                  sample       15   1336.272 ±    0.096    B/op
TracerBenchmarks.newChildWithSpanInScope_extra:newChildWithSpanInScope_extra·p0.999                           sample              20.769              us/op
TracerBenchmarks.newChildWithSpanInScope_extra:·gc.alloc.rate.norm                                            sample       15   1712.572 ±   37.532    B/op
TracerBenchmarks.newChildWithSpanInScope_unsampled:newChildWithSpanInScope_unsampled·p0.999                   sample               1.682              us/op
TracerBenchmarks.newChildWithSpanInScope_unsampled:·gc.alloc.rate.norm                                        sample       15     96.010 ±    0.003    B/op
TracerBenchmarks.newChildWithSpanInScope_unsampled_extra:newChildWithSpanInScope_unsampled_extra·p0.999       sample              16.352              us/op
TracerBenchmarks.newChildWithSpanInScope_unsampled_extra:·gc.alloc.rate.norm                                  sample       15    424.346 ±    0.046    B/op
TracerBenchmarks.nextWithSpanInScope:nextWithSpanInScope·p0.999                                               sample              11.152              us/op
TracerBenchmarks.nextWithSpanInScope:·gc.alloc.rate.norm                                                      sample       15   1384.177 ±   37.565    B/op
TracerBenchmarks.nextWithSpanInScope_extra:nextWithSpanInScope_extra·p0.999                                   sample              14.320              us/op
TracerBenchmarks.nextWithSpanInScope_extra:·gc.alloc.rate.norm                                                sample       15   1760.513 ±    0.033    B/op
TracerBenchmarks.nextWithSpanInScope_unsampled:nextWithSpanInScope_unsampled·p0.999                           sample               1.756              us/op
TracerBenchmarks.nextWithSpanInScope_unsampled:·gc.alloc.rate.norm                                            sample       15     96.007 ±    0.001    B/op
TracerBenchmarks.nextWithSpanInScope_unsampled_extra:nextWithSpanInScope_unsampled_extra·p0.999               sample              13.182              us/op
TracerBenchmarks.nextWithSpanInScope_unsampled_extra:·gc.alloc.rate.norm                                      sample       15    565.759 ±    4.174    B/op
TracerBenchmarks.startScopedSpanWithParent:startScopedSpanWithParent·p0.999                                   sample              11.742              us/op
TracerBenchmarks.startScopedSpanWithParent:·gc.alloc.rate.norm                                                sample       15   1296.177 ±    0.031    B/op
TracerBenchmarks.startScopedSpanWithParent_extra:startScopedSpanWithParent_extra·p0.999                       sample              15.318              us/op
TracerBenchmarks.startScopedSpanWithParent_extra:·gc.alloc.rate.norm                                          sample       15   1672.493 ±   37.556    B/op
TracerBenchmarks.startScopedSpanWithParent_unsampled:startScopedSpanWithParent_unsampled·p0.999               sample               1.746              us/op
TracerBenchmarks.startScopedSpanWithParent_unsampled:·gc.alloc.rate.norm                                      sample       15    104.010 ±    0.006    B/op
TracerBenchmarks.startScopedSpanWithParent_unsampled_extra:startScopedSpanWithParent_unsampled_extra·p0.999   sample              17.600              us/op
TracerBenchmarks.startScopedSpanWithParent_unsampled_extra:·gc.alloc.rate.norm                                sample       15    440.389 ±   12.483    B/op
```


after:
```
Benchmark                                                                                                       Mode      Cnt      Score      Error   Units
TracerBenchmarks.joinWithSpanInScope:joinWithSpanInScope·p0.999                                               sample              23.616             us/op
TracerBenchmarks.joinWithSpanInScope:·gc.alloc.rate.norm                                                      sample       15    735.714 ±   4.159    B/op
TracerBenchmarks.joinWithSpanInScope_extra:joinWithSpanInScope_extra·p0.999                                   sample              23.392             us/op
TracerBenchmarks.joinWithSpanInScope_extra:·gc.alloc.rate.norm                                                sample       15    722.041 ±   3.148    B/op
TracerBenchmarks.joinWithSpanInScope_unsampled:joinWithSpanInScope_unsampled·p0.999                           sample               2.274             us/op
TracerBenchmarks.joinWithSpanInScope_unsampled:·gc.alloc.rate.norm                                            sample       15     40.009 ±   0.006    B/op
TracerBenchmarks.joinWithSpanInScope_unsampled_extra:joinWithSpanInScope_unsampled_extra·p0.999               sample              25.632             us/op
TracerBenchmarks.joinWithSpanInScope_unsampled_extra:·gc.alloc.rate.norm                                      sample       15     40.053 ±   0.004    B/op
TracerBenchmarks.newChildWithSpanInScope:newChildWithSpanInScope·p0.999                                       sample              11.216             us/op
TracerBenchmarks.newChildWithSpanInScope:·gc.alloc.rate.norm                                                  sample       15   1288.179 ±  75.129    B/op
TracerBenchmarks.newChildWithSpanInScope_extra:newChildWithSpanInScope_extra·p0.999                           sample              12.048             us/op
TracerBenchmarks.newChildWithSpanInScope_extra:·gc.alloc.rate.norm                                            sample       15   1512.306 ±  12.507    B/op
TracerBenchmarks.newChildWithSpanInScope_unsampled:newChildWithSpanInScope_unsampled·p0.999                   sample               0.779             us/op
TracerBenchmarks.newChildWithSpanInScope_unsampled:·gc.alloc.rate.norm                                        sample       15     96.007 ±   0.002    B/op
TracerBenchmarks.newChildWithSpanInScope_unsampled_extra:newChildWithSpanInScope_unsampled_extra·p0.999       sample               9.920             us/op
TracerBenchmarks.newChildWithSpanInScope_unsampled_extra:·gc.alloc.rate.norm                                  sample       15    336.123 ±   0.008    B/op
TracerBenchmarks.nextWithSpanInScope:nextWithSpanInScope·p0.999                                               sample              11.040             us/op
TracerBenchmarks.nextWithSpanInScope:·gc.alloc.rate.norm                                                      sample       15   1336.182 ±   0.012    B/op
TracerBenchmarks.nextWithSpanInScope_extra:nextWithSpanInScope_extra·p0.999                                   sample              11.808             us/op
TracerBenchmarks.nextWithSpanInScope_extra:·gc.alloc.rate.norm                                                sample       15   1544.300 ±  45.141    B/op
TracerBenchmarks.nextWithSpanInScope_unsampled:nextWithSpanInScope_unsampled·p0.999                           sample               0.816             us/op
TracerBenchmarks.nextWithSpanInScope_unsampled:·gc.alloc.rate.norm                                            sample       15     96.006 ±   0.001    B/op
TracerBenchmarks.nextWithSpanInScope_unsampled_extra:nextWithSpanInScope_unsampled_extra·p0.999               sample              16.866             us/op
TracerBenchmarks.nextWithSpanInScope_unsampled_extra:·gc.alloc.rate.norm                                      sample       15    448.328 ±  12.505    B/op
TracerBenchmarks.startScopedSpanWithParent:startScopedSpanWithParent·p0.999                                   sample               9.600             us/op
TracerBenchmarks.startScopedSpanWithParent:·gc.alloc.rate.norm                                                sample       15   1248.145 ±  37.564    B/op
TracerBenchmarks.startScopedSpanWithParent_extra:startScopedSpanWithParent_extra·p0.999                       sample              10.336             us/op
TracerBenchmarks.startScopedSpanWithParent_extra:·gc.alloc.rate.norm                                          sample       15   1496.261 ±  50.079    B/op
TracerBenchmarks.startScopedSpanWithParent_unsampled:startScopedSpanWithParent_unsampled·p0.999               sample               0.660             us/op
TracerBenchmarks.startScopedSpanWithParent_unsampled:·gc.alloc.rate.norm                                      sample       15    104.006 ±   0.001    B/op
TracerBenchmarks.startScopedSpanWithParent_unsampled_extra:startScopedSpanWithParent_unsampled_extra·p0.999   sample               5.975             us/op
TracerBenchmarks.startScopedSpanWithParent_unsampled_extra:·gc.alloc.rate.norm                                sample       15    352.117 ±  12.520    B/op
```